requirements.txt
easydict==1.9
imageio==2.13.3
ipdb==0.13.9
PyYAML==6.0
tensorboard==2.12.3
tqdm==4.62.3
trimesh==3.15.8
visdom==0.1.8.9
termcolor==1.1.0
matplotlib==3.5.1
numpy==1.23.5
plyfile==0.7.4
numpy-stl==2.17.1
PyMCubes==0.1.4
pykdtree==1.3.7.post0
Pillow==9.4.0